fold System.IO.FileSystem into System.Private.CoreLib
