+++ src/libraries/System.IO.FileSystem/ref/System.IO.FileSystem.Forwards.cs
@@ -7,3 +7,17 @@
 [assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(Microsoft.Win32.SafeHandles.SafeFileHandle))]
 [assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.IO.FileOptions))]
 [assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.IO.FileStream))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.IO.Directory))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.IO.DirectoryInfo))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.IO.EnumerationOptions))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.IO.File))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.IO.FileInfo))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.IO.FileSystemInfo))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.IO.MatchCasing))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.IO.MatchType))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.IO.SearchOption))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.IO.Enumeration.FileSystemEntry))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.IO.Enumeration.FileSystemEnumerable<>))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.IO.Enumeration.FileSystemEnumerator<>))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.IO.Enumeration.FileSystemName))]
+
