+++ src/libraries/System.Private.CoreLib/src/System/IO/Enumeration/FileSystemEntry.cs
