+++ src/libraries/System.Private.CoreLib/src/System/IO/Enumeration/FileSystemEnumerable.cs
