+++ src/libraries/System.Private.CoreLib/src/System/IO/Enumeration/FileSystemEnumerator.cs
