+++ src/libraries/System.Private.CoreLib/src/System/IO/EnumerationOptions.cs
