+++ src/libraries/System.Private.CoreLib/src/System/IO/File.cs
@@ -352,7 +352,11 @@ namespace System.IO
                 {
                     int n = fs.Read(bytes, index, count);
                     if (n == 0)
+#if !MS_IO_REDIST
+                        ThrowHelper.ThrowEndOfFileException();
+#else
                         throw Error.GetEndOfFile();
+#endif
                     index += n;
                     count -= n;
                 }
@@ -815,7 +819,11 @@ namespace System.IO
 #endif
                     if (n == 0)
                     {
+#if !MS_IO_REDIST
+                        ThrowHelper.ThrowEndOfFileException();
+#else
                         throw Error.GetEndOfFile();
+#endif
                     }
 
                     index += n;
