+++ src/libraries/System.Private.CoreLib/src/System/IO/FileInfo.cs
