+++ src/libraries/System.Private.CoreLib/src/System/IO/FileSystem.Win32.cs
