+++ src/libraries/System.Private.CoreLib/src/System/IO/FileSystem.Windows.cs
