+++ src/libraries/System.Private.CoreLib/src/System/IO/FileSystemInfo.Windows.cs
