+++ src/libraries/System.Private.CoreLib/src/System/IO/FileSystemInfo.cs
