+++ src/libraries/System.Private.CoreLib/src/System/IO/Iterator.cs
