+++ src/libraries/System.Private.CoreLib/src/System/IO/MatchCasing.cs
