+++ src/libraries/System.Private.CoreLib/src/System/IO/MatchType.cs
