+++ src/libraries/System.Private.CoreLib/src/System/IO/ReadLinesIterator.cs
