+++ src/libraries/System.Private.CoreLib/src/System/IO/SearchOption.cs
