+++ src/libraries/System.Private.CoreLib/src/System/IO/SearchTarget.cs
