+++ src/libraries/System.Private.CoreLib/src/System/Resources/FileBasedResourceGroveler.cs
@@ -15,8 +15,6 @@ using System.Diagnostics;
 using System.Globalization;
 using System.IO;
 
-using Internal.IO;
-
 namespace System.Resources
 {
     internal sealed class FileBasedResourceGroveler : IResourceGroveler
